// stundenplaner-ui/src/app/enums/days.ts
export enum Days {
    MONDAY_MORNING = "MONDAY_MORNING",
    MONDAY_AFTERNOON = "MONDAY_AFTERNOON",
    TUESDAY_MORNING = "TUESDAY_MORNING",
    TUESDAY_AFTERNOON = "TUESDAY_AFTERNOON",
    WEDNESDAY_MORNING = "WEDNESDAY_MORNING",
    WEDNESDAY_AFTERNOON = "WEDNESDAY_AFTERNOON",
    THURSDAY_MORNING = "THURSDAY_MORNING",
    THURSDAY_AFTERNOON = "THURSDAY_AFTERNOON",
    FRIDAY_MORNING = "FRIDAY_MORNING",
    FRIDAY_AFTERNOON = "FRIDAY_AFTERNOON",
}